Fix brokean protobuf-native schema (#8621)
Motivation

The protobuf-native schema was broken due to two concurrent merges
